Final Version

1. 修复初始化项目时添加联系人失败
2. 修复文件IO下，id与数据库同步的问题
3. 修复空列表情况下退出抛出异常
4. 修复数据库模式下导入数据时，选择覆盖原数据数据库修改不同步
5. 修复GUI删除前不询问
6. 修复GUI退出时没有保存数据到文件

+++ view/gui.py
@@ -329,6 +329,9 @@ class InfoFrame(tk.Frame):
 
     # 删除联系人事件
     def delete(self):
+        confirm = askyesno(title='删除', message='确定删除')
+        if not confirm:
+            return
         if controller.delete(DataSet.data, self.data.id):
             DataSet.change = True
             showinfo(title='删除', message='删除成功')
@@ -407,3 +410,5 @@ def gui():
     s.place(relx=0, rely=0, relwidth=1, relheight=0.1)
 
     root.mainloop()
+
+    controller.save(data)
